minor fixes and refactoring
